// UD4/actividades/misrecetas/src/app/models/recipe.model.ts
export interface Recipes {
  id: number;
  name: string;
  ingredients: string[];
  instructions: string[];
  prepTimeMinutes: number;
  cookTimeMinutes: number;
  servings: number;
  difficulty: string;
  cuisine: string;
  caloriesPerServing: number;
  tags: string[];
  userId: number;
  image: string;
  rating: number;
  reviewCount: number;
  mealType: string[];
}

export interface RecipesResponse {
  recipes: Recipes[];
  total: number;
  skip: number;
  limit: number;
}

export interface RecipeFormData {
  name: string;
  difficulty: string;
  prepTimeMinutes: number;
  cookTimeMinutes: number;
  ingredients: string;
  instructions: string;
  servings?: number;
  tags?: string;
}